Rework the default payload available in the heartbeat feature to make it testable (and, it turns out, more extensible and maintainable)
- Separate 'generic' from 'azure-specific' payload items

- Provide a way to add further default payload items in an easy way

- Provide a way to test the default payloads robustly (no timers/waits during unit tests)
