+++ Test/Microsoft.ApplicationInsights.Test/Shared/Extensibility/Implementation/Tracing/HeartbeatTests.cs
@@ -4,9 +4,9 @@
     using System.Collections.Generic;
     using System.Globalization;
     using System.Linq;
-    using System.Threading.Tasks;
     using Microsoft.ApplicationInsights.DataContracts;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
+    using Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing.Mocks;
 
     [TestClass]
     public class HeartbeatTests
@@ -234,18 +234,21 @@
         [TestMethod]
         public void HeartbeatPayloadContainsOnlyAllowedDefaultPayloadFields()
         {
-            var allDefaultFields = AzureMetadataRequest.DefaultFields.Union(HeartbeatDefaultPayload.AllDefaultFields);
+            var assemblyDefFields = new BaseHeartbeatProperties();
+            var azureInstanceDefFields = new AzureHeartbeatProperties();
+
+            var allDefaultFields = azureInstanceDefFields.DefaultFields.Union(assemblyDefFields.DefaultFields).ToList();
 
             using (var hbeat = new HeartbeatProvider())
             {
                 List<string> disableHbProps = new List<string>();
 
-                for (int i = 0; i < HeartbeatDefaultPayload.DefaultFields.Length; ++i)
+                for (int i = 0; i < allDefaultFields.Count(); ++i)
                 {
                     if (i % 2 == 0)
                     {
-                        disableHbProps.Add(HeartbeatDefaultPayload.DefaultFields[i]);
-                        hbeat.ExcludedHeartbeatProperties.Add(HeartbeatDefaultPayload.DefaultFields[i]);
+                        disableHbProps.Add(allDefaultFields[i]);
+                        hbeat.ExcludedHeartbeatProperties.Add(allDefaultFields[i]);
                     }
                 }
 
@@ -309,25 +312,16 @@
         [TestMethod]
         public void SentHeartbeatContainsExpectedDefaultFields()
         {
-            using (var hbeat = new HeartbeatProvider())
+            using (var hbeatMock = new HeartbeatProviderMock())
             {
-                hbeat.Initialize(configuration: null);
+                var baseHbeatProps = new BaseHeartbeatProperties().DefaultFields;
 
-                var msg = (MetricTelemetry)hbeat.GatherData();
-                Assert.IsNotNull(msg);
+                var taskWaiter = HeartbeatDefaultPayload.PopulateDefaultPayload(new string[] { }, hbeatMock, false).ConfigureAwait(false);
+                Assert.IsTrue(taskWaiter.GetAwaiter().GetResult()); // no await for tests
 
-                foreach (string field in HeartbeatDefaultPayload.DefaultFields)
+                foreach (string fieldName in baseHbeatProps)
                 {
-                    try
-                    {
-                        var fieldPayload = msg.Properties.Single(a => string.Compare(a.Key, field) == 0);
-                        Assert.IsNotNull(fieldPayload);
-                        Assert.IsFalse(string.IsNullOrEmpty(fieldPayload.Value));
-                    }
-                    catch (Exception)
-                    {
-                        Assert.Fail(string.Format(CultureInfo.CurrentCulture, "The default field '{0}' is not present exactly once in a sent message.", field));
-                    }
+                    Assert.IsTrue(hbeatMock.HeartbeatProperties.ContainsKey(fieldName));
                 }
             }
         }
@@ -363,8 +357,10 @@
             using (var hbeat = new HeartbeatProvider())
             {
                 hbeat.Initialize(configuration: null);
-
-                foreach (string key in HeartbeatDefaultPayload.DefaultFields)
+                BaseHeartbeatProperties defFields = new BaseHeartbeatProperties();
+                AzureHeartbeatProperties azFields = new AzureHeartbeatProperties();
+                var defaultFields = defFields.DefaultFields.Union(azFields.DefaultFields).ToList();
+                foreach (string key in defaultFields)
                 {
                     Assert.IsFalse(hbeat.SetHeartbeatProperty(key, "test", true));
                 }
@@ -377,8 +373,10 @@
             using (var hbeat = new HeartbeatProvider())
             {
                 hbeat.Initialize(configuration: null);
-
-                foreach (string key in HeartbeatDefaultPayload.DefaultFields)
+                BaseHeartbeatProperties defFields = new BaseHeartbeatProperties();
+                AzureHeartbeatProperties azFields = new AzureHeartbeatProperties();
+                var defaultFields = defFields.DefaultFields.Union(azFields.DefaultFields).ToList();
+                foreach (string key in defaultFields)
                 {
                     Assert.IsFalse(hbeat.AddHeartbeatProperty(key, "test", true));
                 }
@@ -388,15 +386,10 @@
         [TestMethod]
         public void EnsureAllTargetFrameworksRepresented()
         {
-            using (var hbeat = new HeartbeatProvider())
-            {
-                hbeat.Initialize(configuration: null);
-
-                var msg = (MetricTelemetry)hbeat.GatherData();
-
-                Assert.IsTrue(msg.Properties.ContainsKey("baseSdkTargetFramework"));
-                Assert.IsFalse(msg.Properties["baseSdkTargetFramework"].Equals("undefined", StringComparison.OrdinalIgnoreCase));
-            }
+            BaseHeartbeatProperties bp = new BaseHeartbeatProperties();
+            var noAwaiterForTests = bp.SetDefaultPayload(new string[] { }, new HeartbeatProviderMock()).ConfigureAwait(false);
+            Assert.IsTrue(noAwaiterForTests.GetAwaiter().GetResult());
+            // this is enough to ensure we've hit all of our target framework paths (unless a new one has been added, the purpose of this test)
         }
 
         [TestMethod]
@@ -438,135 +431,26 @@
             }
         }
 
-        private class TestRequestorStub : IAzureMetadataRequestor
-        {
-            private Func<IEnumerable<string>> GetAllFieldsFunc = null;
-            private Func<string, string> GetSingleFieldFunc = null;
-
-            public TestRequestorStub(Func<IEnumerable<string>> getAllFields = null, Func<string,string> getSingleFieldFunc = null)
-            {
-                this.GetAllFieldsFunc = getAllFields;
-                if (getAllFields == null)
-                {
-                    this.GetAllFieldsFunc = this.GetAllFields;
-                }
-
-                this.GetSingleFieldFunc = getSingleFieldFunc;
-                if (getSingleFieldFunc == null)
-                {
-                    this.GetSingleFieldFunc = this.GetSingleField;
-                }
-            }
-
-            public Dictionary<string, string> computeFields = new Dictionary<string, string>();
-
-            public Task<string> GetAzureComputeMetadata(string fieldName)
-            {
-                return Task.FromResult(this.GetSingleFieldFunc(fieldName));
-            }
-
-            private string GetSingleField(string fieldName)
-            {
-                if (this.computeFields.ContainsKey(fieldName))
-                {
-                    return this.computeFields[fieldName];
-                }
-
-                return string.Empty;
-            }
-
-            public Task<IEnumerable<string>> GetAzureInstanceMetadataComputeFields()
-            {
-                return Task.FromResult(this.GetAllFieldsFunc());
-            }
-
-            private IEnumerable<string> GetAllFields()
-            {
-                IEnumerable<string> fields = this.computeFields.Keys.ToArray();
-                return fields;
-            }
-
-        }
-
-        private class TestHeartbeatProviderStub : IHeartbeatProvider
-        {
-            public Dictionary<string, HeartbeatPropertyPayload> HeartbeatProperties = new Dictionary<string, HeartbeatPropertyPayload>();
-            public List<string> ExcludedPropertyFields = new List<string>();
-
-            public string InstrumentationKey { get; set; }
-
-            public bool IsHeartbeatEnabled { get; set; }
-
-            public bool EnableInstanceMetadata { get; set; }
-
-            public TimeSpan HeartbeatInterval { get; set; }
-
-            public IList<string> ExcludedHeartbeatProperties { get => this.ExcludedPropertyFields; }
-
-            public TestHeartbeatProviderStub()
-            {
-                this.InstrumentationKey = Guid.NewGuid().ToString();
-                this.IsHeartbeatEnabled = true;
-                this.EnableInstanceMetadata = true;
-                this.HeartbeatInterval = TimeSpan.FromSeconds(31);
-            }
-
-            public bool AddHeartbeatProperty(string propertyName, bool overrideDefaultField, string propertyValue, bool isHealthy)
-            {
-                this.HeartbeatProperties.Add(
-                    propertyName, 
-                    new HeartbeatPropertyPayload()
-                    {
-                        IsHealthy = isHealthy,
-                        IsUpdated = true,
-                        PayloadValue = propertyValue
-                    });
-
-                return true;
-            }
-
-            public void Dispose()
-            {
-            }
-
-            public void Initialize(TelemetryConfiguration configuration)
-            {
-            }
-
-            public bool SetHeartbeatProperty(string propertyName, bool overrideDefaultField, string propertyValue = null, bool? isHealthy = null)
-            {
-                if (this.HeartbeatProperties.ContainsKey(propertyName))
-                {
-                    HeartbeatPropertyPayload pl = this.HeartbeatProperties[propertyName];
-                    pl.IsHealthy = isHealthy.GetValueOrDefault(pl.IsHealthy);
-                    pl.PayloadValue = propertyValue ?? pl.PayloadValue;
-                    pl.IsUpdated = true;
-
-                    return true;
-                }
-                return false;
-            }
-        }
-
         [TestMethod]
         public void GetAzureInstanceMetadataFieldsAsExpected()
         {
-            using (var hbeat = new TestHeartbeatProviderStub())
+            using (var hbeatMock = new HeartbeatProviderMock())
             {
-                TestRequestorStub myRequestor = new TestRequestorStub();
+                AzureInstanceMetadataRequestMock azureInstanceRequestorMock = new AzureInstanceMetadataRequestMock();
+                AzureHeartbeatProperties azFields = new AzureHeartbeatProperties(azureInstanceRequestorMock);
                 int counter = 1;
-                foreach (string field in HeartbeatDefaultPayload.DefaultOptionalFields)
+                foreach (string field in azFields.DefaultFields)
                 {
-                    myRequestor.computeFields.Add(field, $"testValue{counter++}");
+                    azureInstanceRequestorMock.computeFields.Add(field, $"testValue{counter++}");
                 }
 
-                HeartbeatDefaultPayload.PopulateDefaultPayload(new string[] { }, hbeat, myRequestor);
-                // this is an async call, and will most defintely finish after the next call. Instead
-                // of just doing a timeout and allowing for uncertainty in tests, let's be a bit more
-                
-                foreach (string fieldName in HeartbeatDefaultPayload.DefaultOptionalFields)
+                var taskWaiter = azFields.SetDefaultPayload(new string[] { }, hbeatMock).ConfigureAwait(false);
+                Assert.IsTrue(taskWaiter.GetAwaiter().GetResult()); // no await for tests
+
+                foreach (string fieldName in azFields.DefaultFields)
                 {
-                    Assert.IsTrue(hbeat.HeartbeatProperties.ContainsKey(fieldName));
+                    Assert.IsTrue(hbeatMock.HeartbeatProperties.ContainsKey(fieldName));
+                    Assert.IsFalse(string.IsNullOrEmpty(hbeatMock.HeartbeatProperties[fieldName].PayloadValue));
                 }
             }
         }
@@ -574,15 +458,20 @@
         [TestMethod]
         public void FailToObtainAzureInstanceMetadataFieldsAltogether()
         {
-            using (var hbeat = new TestHeartbeatProviderStub())
+            using (var hbeatMock = new HeartbeatProviderMock())
             {
-                TestRequestorStub myRequestor = new TestRequestorStub();
+                AzureInstanceMetadataRequestMock azureInstanceRequestorMock = new AzureInstanceMetadataRequestMock();
+                var azFields = new AzureHeartbeatProperties(azureInstanceRequestorMock);
+                var defaultFields = azFields.DefaultFields;
+                // not adding the fields we're looking for, simulation of the Azure Instance Metadata service not being present...
 
-                HeartbeatDefaultPayload.PopulateDefaultPayload(new string[] { }, hbeat, myRequestor);
+                var taskWaiter = azFields.SetDefaultPayload(new string[] { }, hbeatMock).ConfigureAwait(false);
+                Assert.IsTrue(taskWaiter.GetAwaiter().GetResult()); // nop await for tests
 
-                foreach (string fieldName in HeartbeatDefaultPayload.DefaultOptionalFields)
+                foreach (string fieldName in defaultFields)
                 {
-                    Assert.IsFalse(hbeat.HeartbeatProperties.ContainsKey(fieldName));
+                    Assert.IsTrue(hbeatMock.HeartbeatProperties.ContainsKey(fieldName));
+                    Assert.IsTrue(string.IsNullOrEmpty(hbeatMock.HeartbeatProperties[fieldName].PayloadValue));
                 }
             }
         }
