+++ src/Microsoft.ApplicationInsights/Managed/Shared/Extensibility/Implementation/Tracing/AzureHeartbeatProperties.cs
@@ -1,0 +1,81 @@
+﻿namespace Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Threading.Tasks;
+
+    internal class AzureHeartbeatProperties : IHeartbeatDefaultPayloadProvider
+    {
+        // internal for testing
+        internal readonly List<string> DefaultFields = new List<string>()
+        {
+            "osType",
+            "location",
+            "name",
+            "offer",
+            "platformFaultDomain",
+            "platformUpdateDomain",
+            "publisher",
+            "sku",
+            "version",
+            "vmId",
+            "vmSize"
+        };
+
+        /// <summary>
+        /// Flags that will tell us whether or not Azure VM metadata has been attempted to be gathered or not, and
+        /// if we should even attempt to look for it in the first place. Static to ensure we only ever attempt this
+        /// once for any given consuming application.
+        /// </summary>
+        private static bool isAzureMetadataCheckCompleted = false;
+
+        private IAzureMetadataRequestor azureInstanceMetadataRequestor = null;
+
+        /// <summary>
+        /// Constructor for the Azure specific fields to inject into the heartbeat payload
+        /// </summary>
+        /// <param name="azInstanceMetadataHandler">(for testing) Azure metadata request handler to use when requesting data from azure specifically. If left as null, an instance of AzureMetadatRequestor is used.</param>
+        public AzureHeartbeatProperties(IAzureMetadataRequestor azInstanceMetadataHandler = null)
+        {
+            this.azureInstanceMetadataRequestor = azInstanceMetadataHandler;
+            if (this.azureInstanceMetadataRequestor == null)
+            {
+                this.azureInstanceMetadataRequestor = new AzureMetadataRequestor();
+            }
+        }
+
+        public bool IsKeyword(string keyword)
+        {
+            return this.DefaultFields.Contains(keyword, StringComparer.OrdinalIgnoreCase);
+        }
+
+        public async Task<bool> SetDefaultPayload(IEnumerable<string> disabledFields, IHeartbeatProvider provider)
+        {
+            bool hasSetFields = false;
+
+            if (!AzureHeartbeatProperties.isAzureMetadataCheckCompleted)
+            {
+                // only ever attempt this once
+                AzureHeartbeatProperties.isAzureMetadataCheckCompleted = true;
+
+                var allFields = await this.azureInstanceMetadataRequestor.GetAzureInstanceMetadataComputeFields()
+                                .ConfigureAwait(false);
+
+                var enabledImdsFields = this.DefaultFields.Except(disabledFields);
+                foreach (string field in enabledImdsFields)
+                {
+                    provider.AddHeartbeatProperty(
+                        propertyName: field,
+                        overrideDefaultField: true,
+                        propertyValue: await this.azureInstanceMetadataRequestor.GetAzureComputeMetadata(field)
+                            .ConfigureAwait(false),
+                        isHealthy: true);
+                    hasSetFields = true;
+                }
+            }
+
+            return hasSetFields;
+        }
+    }
+}
